+++ src/Events/Integrations/Plugins/Elementor/Widgets/Event_Export.php
@@ -126,17 +126,6 @@ class Event_Export extends Abstract_Widget {
 	}
 
 	/**
-	 * Get the template args for the widget preview.
-	 *
-	 * @since TBD
-	 *
-	 * @return array The template args for the preview.
-	 */
-	protected function preview_args(): array {
-		return $this->template_args();
-	}
-
-	/**
 	 * Get the class used for the dropdown.
 	 *
 	 * @since TBD
